format and lint
